Add Haskell 014 and fix Rust 014

+++ src/014/p014.rs
@@ -13,7 +13,7 @@ fn generate_collatz_sequence_starting_at(number: i64) -> Vec<i64> {
     return numbers;
 }
 
-fn solve() -> i32 {
+fn solve() -> i64 {
     let mut longest_chain = Vec::new();
     let mut number = 1;
     let mut highest_number = 1;
